refactor(dccpeep): extract board/game passes into peep_pass_minmax.c
Move the 14 application-specific ttt/chess passes (pass_posfunc_b_cache, the pass_minmax_* family, pass_reuse_board_addr_for_zero_store, pass_winner_check_dec_a, pass_global_board_const_offsets) plus their private helpers (peep_is_pos_func_label, pack_str_replace) into peep_pass_minmax.c. Promote the genuinely shared helpers peep_in_function_range, peep_range_has_debug_annotations, bc_regalloc_claimed_before, and stride_parse_ld_r_ix_neg to the internal header.

dccpeep.c drops from ~10.6k to ~9.2k lines. Validation: strict build clean; all 316 raw inputs byte-identical in default, -Os, and -fundocumented-z80 modes; direct fixtures pass.
